requirements.txt
Django==5.2
psycopg2==2.9.10
python-decouple==3.8
requests==2.32.3
djangorestframework==3.16.0